lyra/requirements.txt
pandas==1.1.0
matplotlib==3.3.0
tenacity==6.2.0
altair==4.1.0
scipy==1.5.2
geopandas==0.8.1
topojson==1.0rc10
descartes==1.1.0
shapely==1.7.0
celery==4.4.7
jinja2==2.11.2
requests==2.24.0
redis==3.5.3
pyyaml==5.3.1
fastapi==0.60.1
pydantic==1.6.1
orjson==3.3.0
pint==0.14
aiofiles==0.5.0
rtree==0.9.4
aiohttp==3.6.2
typing-extensions==3.7.4.2
sqlalchemy==1.3.18
azure-storage-file-share==12.1.2
pyodbc==4.0.30
